+++ integrations/openapi/src/components/configureSource.tsx
@@ -1,27 +1,35 @@
-import { createComponent, ExposableError, InstallationConfigurationProps } from '@gitbook/runtime';
+import { ConfigureContentSourceProps, createComponent, ExposableError } from '@gitbook/runtime';
 
-import type { OpenAPIRuntimeEnvironment, OpenAPIRuntimeContext } from '../types';
+import type { OpenAPIRuntimeContext } from '../types';
+import type { OpenAPIContentSource } from '../contentSources';
 
 /**
  * ContentKit component to configure the content source.
  */
 export const configureComponent = createComponent<
-    InstallationConfigurationProps<OpenAPIRuntimeEnvironment>,
+    // @ts-expect-error incompatible with PlainObject
+    ConfigureContentSourceProps<OpenAPIContentSource>,
     {
         spec: string | null;
         models: boolean;
     },
-    { action: 'submit' } | { action: 'selectSpec'; spec: string },
+    { action: 'submit' },
     OpenAPIRuntimeContext
 >({
     componentId: 'configureSource',
-    initialState: (props, _, context) => {
+    initialState: (props) => {
+        if (props.contentSource) {
+            return {
+                spec: props.contentSource.dependencies.spec.ref.spec,
+                models: props.contentSource.props.models,
+            };
+        }
         return {
             spec: null,
             models: true,
         };
     },
-    action: async (element, action, ctx) => {
+    action: async (element, action, _ctx) => {
         if (action.action === 'submit') {
             if (!element.state.spec) {
                 throw new ExposableError('Invalid spec URL');
@@ -45,20 +53,9 @@ export const configureComponent = createComponent<
             };
         }
 
-        if (action.action === 'selectSpec') {
-            return {
-                ...element,
-                state: {
-                    ...element.state,
-                    spec: action.spec,
-                },
-            };
-        }
-
         return element;
     },
     render: async (element, context) => {
-        const { api } = context;
         const { installation } = context.environment;
         const { state } = element;
 
@@ -66,31 +63,14 @@ export const configureComponent = createComponent<
             throw new ExposableError('Installation not found');
         }
 
-        const {
-            data: { items: specs },
-        } = await api.orgs.listOpenApiSpecs(installation.target.organization);
         return (
             <configuration>
                 <input
                     label="OpenAPI Specification"
                     hint="Choose the OpenAPI specification to use."
-                    element={
-                        <select
-                            state="spec"
-                            options={specs.map((spec) => ({
-                                label: spec.slug,
-                                id: spec.slug,
-                            }))}
-                            onValueChange={{
-                                action: 'selectSpec',
-                                spec: element.dynamicState('spec'),
-                            }}
-                        />
-                    }
+                    element={<select state="spec" options={{ source: 'openapi' }} />}
                 />
-
                 <divider />
-
                 <input
                     label="Generate models"
                     hint="Generate a models page for all schema components."
@@ -98,7 +78,7 @@ export const configureComponent = createComponent<
                 />
                 <button
                     style="primary"
-                    label="Continue"
+                    label={element.props.submitLabel ?? 'Continue'}
                     disabled={!state.spec}
                     onPress={{
                         action: 'submit',