+++ integrations/openapi/src/parser/revision.ts
@@ -1,52 +1,34 @@
-import type { ContentComputeRevisionEventResponse, InputPage } from '@gitbook/api';
-import type { GenerateGroupPageProps } from '../contentSources';
-import { divideOpenAPISpecSchema, type OpenAPISpecContent } from './spec';
-import { OpenAPIV3 } from '@gitbook/openapi-parser';
+import type { ContentComputeRevisionEventResponse, InputPageDocument } from '@gitbook/api';
+import {
+    extractPageOperations,
+    getOpenAPITree,
+    type OpenAPIPage,
+    type OpenAPISpecContent,
+} from './spec';
 
 /**
  * Generate the revision for the OpenAPI specification.
  */
-export function getRevisionFromSpec(args: {
+export async function getRevisionFromSpec(args: {
     specContent: OpenAPISpecContent;
     props?: {
         models?: boolean;
     };
-}): ContentComputeRevisionEventResponse {
+}): Promise<ContentComputeRevisionEventResponse> {
     const { specContent, props } = args;
-    const groups = divideOpenAPISpecSchema(specContent.schema);
-
-    const groupPages = groups.map((group) => {
-        const documentProps: GenerateGroupPageProps = {
-            doc: 'operations' as const,
-            group: group.id,
-        };
-
-        const page: InputPage = {
-            type: 'document',
-            title: (group.tag ? getTagTitle(group.tag) : '') || improveTagName(group.id),
-            icon: group.tag?.['x-page-icon'],
-            description: group.tag?.['x-page-description'] ?? '',
-            computed: {
-                integration: 'openapi',
-                source: 'generate',
-                props: documentProps,
-                dependencies: {
-                    spec: {
-                        ref: { kind: 'openapi' as const, spec: specContent.slug },
-                    },
-                },
-            },
-        };
-
-        return page;
-    });
+    const tree = getOpenAPITree(specContent.schema);
+    const [tagPages, modelsId] = await Promise.all([
+        openAPIPagesToInputPages(tree, specContent),
+        sha1('models'),
+    ]);
 
     return {
         pages: [
-            ...groupPages,
+            ...tagPages,
             ...(props?.models
                 ? [
                       {
+                          id: modelsId,
                           type: 'document' as const,
                           title: 'Models',
                           computed: {
@@ -66,21 +48,55 @@ export function getRevisionFromSpec(args: {
     };
 }
 
-/**
- * Get the title for a tag.
- */
-function getTagTitle(tag: OpenAPIV3.TagObject) {
-    return tag['x-page-title'] ?? improveTagName(tag.name ?? '');
+async function openAPIPagesToInputPages(
+    pages: OpenAPIPage[],
+    specContent: OpenAPISpecContent,
+): Promise<InputPageDocument[]> {
+    return Promise.all(
+        pages.map(async (openAPIPage) => {
+            const operations = extractPageOperations(openAPIPage);
+
+            const [id, pages] = await Promise.all([
+                sha1(`tag-${openAPIPage.id}`),
+                openAPIPage.pages
+                    ? openAPIPagesToInputPages(openAPIPage.pages, specContent)
+                    : undefined,
+            ]);
+
+            const page: InputPageDocument = {
+                id,
+                type: 'document',
+                title: openAPIPage.title,
+                icon: openAPIPage.tag?.['x-page-icon'],
+                description: openAPIPage.tag?.['x-page-description'] ?? '',
+                pages,
+                computed: operations.length
+                    ? {
+                          integration: 'openapi',
+                          source: 'generate',
+                          props: {
+                              doc: 'operations' as const,
+                              page: openAPIPage.id,
+                          },
+                          dependencies: {
+                              spec: {
+                                  ref: { kind: 'openapi' as const, spec: specContent.slug },
+                              },
+                          },
+                      }
+                    : undefined,
+            };
+
+            return page;
+        }),
+    );
 }
 
-/**
- * Improve a tag name to be used as a page title:
- * - Capitalize the first letter
- * - Split on - and capitalize each word
- */
-function improveTagName(tagName: string) {
-    return tagName
-        .split('-')
-        .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-        .join(' ');
+async function sha1(message: string): Promise<string> {
+    const encoder = new TextEncoder();
+    const data = encoder.encode(message);
+    const hashBuffer = await crypto.subtle.digest('SHA-1', data);
+    const hashArray = Array.from(new Uint8Array(hashBuffer));
+    const hashHex = hashArray.map((b) => b.toString(16).padStart(2, '0')).join('');
+    return hashHex;
 }